Move function from external python file to same python file
